Change getsn() to use size_t for size values.
The size argument and the return value both describe the size of a
memory area, thus size_t is justified.

ok pedro@

+++ sys/lib/libkern/libkern.h
@@ -176,7 +176,7 @@ size_t	 strlcat(char *, const char *, size_t)
 int	 strcmp(const char *, const char *);
 int	 strncmp(const char *, const char *, size_t);
 int	 strncasecmp(const char *, const char *, size_t);
-int	 getsn(char *, int);
+size_t	 getsn(char *, size_t);
 char	*strchr(const char *, int);
 char	*strrchr(const char *, int);
 int	 timingsafe_bcmp(const void *, const void *, size_t);
